backend/services/cluster-inventory/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Kubernetes client
kubernetes>=28.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

